Move compress() out of PointList
